ubah id_barang jadi no_bon, ui terima gadai

+++ database/migrations/2025_03_21_223432_create_transaksi_gadai_table.php
@@ -11,18 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('transaksi_gadai', function (Blueprint $table) {
-        $table->id('id_transaksi');
-        $table->unsignedBigInteger('id_nasabah');
-        $table->unsignedBigInteger('id_barang');
-        $table->date('tanggal_gadai');
-        $table->decimal('jumlah_pinjaman', 15, 2);
-        $table->decimal('bunga', 5, 2);
-        $table->date('jatuh_tempo');
-        $table->foreign('id_nasabah')->references('id_nasabah')->on('nasabah')->onDelete('cascade');
-        $table->foreign('id_barang')->references('id_barang')->on('barang_gadai')->onDelete('cascade');
-        $table->timestamps();
-    });
+        if (!Schema::hasTable('penaksiran_harga')){
+            Schema::create('transaksi_gadai', function (Blueprint $table) {
+            $table->id('id_transaksi');
+            $table->unsignedBigInteger('id_nasabah');
+            $table->unsignedBigInteger('no_bon');
+            $table->date('tanggal_gadai');
+            $table->decimal('jumlah_pinjaman', 15, 2);
+            $table->decimal('bunga', 5, 2);
+            $table->date('jatuh_tempo');
+            $table->foreign('id_nasabah')->references('id_nasabah')->on('nasabah')->onDelete('cascade');
+            $table->foreign('no_bon')->references('no_bon')->on('barang_gadai')->onDelete('cascade');
+            $table->timestamps();
+        });
+
+        }
     }
 
     /**
